+++ CHANGELOG.md
@@ -8,6 +8,9 @@ Please mark backwards incompatible changes with an exclamation mark at the start
 
 ## [Unreleased]
 
+### Added
+- The `Aggregations::DateHistogram` class.
+
 ## [27.4.0] - 2025-04-10
 
 ### Added